[REF] Prepare version 0.1.0
