+++ tests/mfa_random_code_test.rs
@@ -8,13 +8,7 @@ use actix_session::{storage::CookieSessionStore, SessionExt, SessionMiddleware};
 use actix_web::{cookie::Key, get, App, HttpRequest, HttpResponse, HttpServer, Responder};
 use async_trait::async_trait;
 use authfix::{
-    mfa::{HandleMfaRequest, MfaConfig, MfaError},
-    middleware::{AuthMiddleware, PathMatcher},
-    multifactor::random_code_auth::{
-        CodeSendError, CodeSender, MfaRandomCode, RandomCode, MFA_ID_RANDOM_CODE,
-    },
-    session::{config::Routes, handlers::SessionApiHandlers, session_auth::SessionAuthProvider},
-    AuthToken,
+    factor_impl::random_code_auth::{CodeSendError, CodeSender, MfaRandomCodeFactor, RandomCode}, middleware::{AuthMiddleware, PathMatcher}, multifactor::config::{HandleMfaRequest, MfaConfig, MfaError}, session::{config::Routes, handlers::SessionApiHandlers, session_auth::SessionAuthProvider}, AuthToken
 };
 use chrono::{DateTime, Duration, Local, TimeDelta};
 use reqwest::{Client, StatusCode};
@@ -41,7 +35,7 @@ impl HandleMfaRequest for OnlyRandomCodeFactor {
     type User = User;
 
     async fn get_mfa_id_by_user(&self, _: &Self::User) -> Result<Option<String>, MfaError> {
-        Ok(Some(MFA_ID_RANDOM_CODE.to_owned()))
+        Ok(Some(MfaRandomCodeFactor::id().to_owned()))
     }
 }
 
@@ -332,7 +326,7 @@ fn start_test_server(addr: SocketAddr, generator: fn() -> RandomCode) {
                     // Hint:
                     // This is the manual configuration of the auth middleware with a session provider and handlers.
 
-                    let code_factor = Box::new(MfaRandomCode::new(generator, Arc::clone(&sender)));
+                    let code_factor = Box::new(MfaRandomCodeFactor::new(generator, Arc::clone(&sender)));
                     let mfa_config = MfaConfig::new(vec![code_factor], OnlyRandomCodeFactor);
                     let load_user_service = Arc::new(HardCodedLoadUserService);
                     App::new()
