+++ tests/multiple_mfa_test.rs
@@ -3,15 +3,7 @@ use std::{net::SocketAddr, sync::Arc, thread};
 use actix_web::{cookie::Key, get, HttpResponse, HttpServer, Responder};
 use async_trait::async_trait;
 use authfix::{
-    login::{LoadUserByCredentials, LoadUserError, LoginToken},
-    mfa::{HandleMfaRequest, MfaConfig, MfaError},
-    multifactor::{
-        authenticator::{AuthenticatorFactor, MFA_ID_AUTHENTICATOR_TOTP},
-        random_code_auth::{MfaRandomCode, MFA_ID_RANDOM_CODE},
-        Factor,
-    },
-    session::{app_builder::SessionLoginAppBuilder, AccountInfo},
-    AuthToken,
+    factor_impl::{authenticator::AuthenticatorFactor, random_code_auth::MfaRandomCodeFactor}, login::{LoadUserByCredentials, LoadUserError, LoginToken}, multifactor::{config::{HandleMfaRequest, MfaConfig, MfaError}, factor::Factor}, session::{app_builder::SessionLoginAppBuilder, AccountInfo}, AuthToken
 };
 
 use google_authenticator::GoogleAuthenticator;
@@ -49,11 +41,11 @@ impl LoadUserByCredentials for ThreeUserService {
         match login_token.email.as_ref() {
             "joe" => Ok(UserWithMfa {
                 name: "Joe".into(),
-                mfa: Some(MFA_ID_AUTHENTICATOR_TOTP.into()),
+                mfa: Some(AuthenticatorFactor::id().into()),
             }),
             "anna" => Ok(UserWithMfa {
                 name: "anna".into(),
-                mfa: Some(MFA_ID_RANDOM_CODE.into()),
+                mfa: Some(MfaRandomCodeFactor::id().into()),
             }),
             "linda" => Ok(UserWithMfa {
                 name: "linda".into(),
@@ -175,7 +167,7 @@ fn start_test_server(addr: SocketAddr) {
                             Arc::clone(&totp_secret_repo),
                             3,
                         ));
-                    let rand_code: Box<dyn Factor> = Box::new(MfaRandomCode::new(
+                    let rand_code: Box<dyn Factor> = Box::new(MfaRandomCodeFactor::new(
                         single_code_generator,
                         Arc::clone(&sender),
                     ));
