+++ tests/test_utils/mod.rs
@@ -1,11 +1,6 @@
 use async_trait::async_trait;
 use authfix::{
-    login::{LoadUserByCredentials, LoadUserError, LoginToken},
-    multifactor::{
-        random_code_auth::{CodeSendError, CodeSender, RandomCode},
-        GetTotpSecretError, TotpSecretRepository,
-    },
-    session::{AccountInfo, SessionUser},
+    factor_impl::{authenticator::{GetTotpSecretError, TotpSecretRepository}, random_code_auth::{CodeSendError, CodeSender, RandomCode}}, login::{LoadUserByCredentials, LoadUserError, LoginToken}, session::{AccountInfo, SessionUser}
 };
 use chrono::{Local, TimeDelta};
 use serde::{Deserialize, Serialize};
@@ -25,7 +20,7 @@ pub fn test_out_path(path: &str) -> String {
 //
 // For login and mfa tests:
 //
-#[derive(Serialize, Deserialize, Clone)]
+#[derive(Serialize, Deserialize)]
 pub struct User {
     pub email: String,
     pub name: String,
